requirements.txt
pydantic>=2.0
python-dotenv>=1.0
python-docx>=1.0
scipy>=1.10
# Optional (for real LLM providers):
# google-genai>=0.3
# openai>=1.0
# ═══════════════════════════════════════════════
#  iFarma — Генератор синопсиса БЭ-исследования
#  Зависимости проекта
# ═══════════════════════════════════════════════

# --- Ядро ---
requests>=2.31                # HTTP-запросы (Yandex Search, Translate, PubMed)
aiohttp>=3.9                  # Async HTTP (парсинг инструкций vidal/rls/medi/grls)

# --- Научные вычисления ---
numpy>=1.24                   # Массивы — PK-кривая, фарм. расчёты

# --- Генерация документов ---
matplotlib>=3.7               # PK-кривая (Cmax/AUC график)

# --- LLM ---
groq>=0.4                     # Groq API (LLaMA 3.3 / LLaMA 4 Scout)
# google-generativeai>=0.5    # Gemini API (опционально, альтернативный LLM)

# --- RAG (опционально) ---
chromadb>=0.4                 # Векторная БД для RAG-поиска по Решению №85
# sentence-transformers>=2.2  # Эмбеддинги (если не используется Gemini)

# --- Разработка (опционально) ---
# pytest>=7.0
# black>=23.0
# ruff>=0.1
